+++ src/Silex/Provider/ServiceControllerServiceProvider.php
@@ -20,7 +20,7 @@ class ServiceControllerServiceProvider implements ServiceProviderInterface
     public function register(Application $app)
     {
         $app['resolver'] = $app->share($app->extend('resolver', function ($resolver, $app) {
-            return new ServiceControllerResolver($resolver, $app);
+            return new ServiceControllerResolver($resolver, $app['callback_resolver']);
         }));
     }