+++ src/Silex/Application.php
@@ -43,7 +43,7 @@ use Silex\EventListener\StringToResponseListener;
  */
 class Application extends \Pimple implements HttpKernelInterface, TerminableInterface
 {
-    const VERSION = '1.0.2-DEV';
+    const VERSION = '1.0.2';
 
     const EARLY_EVENT = 512;
     const LATE_EVENT  = -512;