+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 * Only convert attributes on the request that actually exist.
+* Add support for using a service method as a converter.
 
 1.1.0 (2013-07-04)
 ------------------