+++ doc/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 * Added run() on Route to be able to define the controller code
 * Deprecated TwigCoreExtension (register the new HttpFragmentServiceProvider instead)
 * Added HttpFragmentServiceProvider
+* Add support for using a service method as a converter.
 
 1.1.1 (2013-10-11)
 ------------------
@@ -14,7 +15,6 @@ Changelog
 * Removed or replaced deprecated Symfony code
 * Updated code to take advantages of 2.3 new features
 * Only convert attributes on the request that actually exist.
-* Add support for using a service method as a converter.
 
 1.1.0 (2013-07-04)
 ------------------