+++ doc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 1.0.2 (2013-XX-XX)
 ------------------
 
-* n/a
+* Fixed SecurityServiceProvider to use null as a fake controller so that routes can be dumped
 
 1.0.1 (2013-07-04)
 ------------------