+++ doc/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.0.2 (2013-XX-XX)
+1.0.2 (2013-10-30)
 ------------------
 
 * Fixed SecurityServiceProvider to use null as a fake controller so that routes can be dumped